refactor(ors_control_app): extract routes/sampling, routes/fleet, lib/state
- routes/sampling.ts: /api/sample-road-points (Overture road sampling +
  bbox/boundary clip + LRU cache lookup)
- routes/fleet.ts: /api/fleet-config*, /api/datasets, /api/datasets/activate
  (with FLEET_CONFIG_SCHEMAS, ORS_PROFILE_TO_VEHICLE_TYPE constants)
- lib/state.ts: getActiveRegionOverride / setActiveRegionOverride encapsulate
  the activeRegionOverride mutable state shared between the regions and
  fleet/datasets routers.

server/index.ts callsites for activeRegionOverride rewritten to the new
getter/setter pair. Lines down 3,001 -> 2,700.

Phase 1 of holistic-control-app-refactor-final, route extraction step 4.
No behaviour change. tsc passes cleanly.
